+++ libs/webb-ui-components/src/components/SideBar/SideBarMenu.tsx
@@ -4,7 +4,7 @@ import * as Dialog from '@radix-ui/react-dialog';
 import { HamburgerMenu, ExternalLinkLine } from '@webb-tools/icons';
 
 import { Typography } from '../../typography/Typography';
-import { SideBarItem } from '../SideBarItem';
+import { SideBarItem } from './Item';
 import { ThemeToggle } from '../ThemeToggle';
 import { Link } from '../Link';
 import { SidebarProps } from './types';
