+++ libs/webb-ui-components/src/components/SideBar/index.ts
@@ -1,4 +1,6 @@
 export * from './SideBar';
 export * from './SideBarMenu';
+export * from './Item';
+export * from './SubItem';
 
 export * from './types';
