+++ libs/webb-ui-components/src/components/SideBar/types.ts
@@ -1,6 +1,5 @@
 import { IconBase } from '@webb-tools/icons/types';
 
-import { SideBarItemProps } from '../SideBarItem';
 import { LogoProps } from '../Logo/types';
 
 export type SideBarFooterProps = {
@@ -17,4 +16,30 @@ export type SidebarProps = {
   items: SideBarItemProps[];
   footer: SideBarFooterProps;
   className?: string;
+};
+
+export type SideBarItemProps = {
+  name: string;
+  isInternal: boolean;
+  href: string;
+  Icon: (props: IconBase) => JSX.Element;
+  subItems: SideBarSubItemProps[];
+};
+
+export type SideBarExtraItemProps = {
+  isSidebarOpen?: boolean;
+  isActive?: boolean;
+  setIsActive?: () => void;
+};
+
+export type SideBarSubItemProps = {
+  name: string;
+  isInternal: boolean;
+  href: string;
+};
+
+export type SideBarExtraSubItemProps = {
+  isActive?: boolean;
+  setItemIsActive?: () => void;
+  setSubItemIsActive?: () => void;
 };
