--- libs/webb-ui-components/src/components/SideBarItem/index.ts
@@ -1,3 +1,0 @@
-export * from './SideBarItem';
-
-export * from './types';
