+++ libs/webb-ui-components/src/components/index.ts
@@ -49,7 +49,6 @@ export * from './Popover';
 export * from './Progress';
 export * from './Radio';
 export * from './SideBar';
-export * from './SideBarItem';
 export * from './Slider';
 export * from './Socials';
 export * from './Stats';
